feature: useRestaurant.js 파일을 생성하여 관심사 분리
