+++ db.json
@@ -35,6 +35,12 @@
       "name": "도스타코스 선릉점",
       "description": "멕시칸 캐주얼 그릴",
       "category": "기타"
+    },
+    {
+      "id": 1731153892123,
+      "name": "test",
+      "description": "test",
+      "category": "중식"
     }
   ]
 }
